+++ Source/SwiftLintFramework/Rules/ValidDocsRule.swift
@@ -51,7 +51,6 @@ func declarationReturns(declaration: String, kind: SwiftDeclarationKind? = nil) 
     guard let outsideBracesMatch = matchOutsideBraces(declaration) else {
         return false
     }
-
     return outsideBracesMatch.containsString("->")
 }
 
@@ -66,6 +65,12 @@ func matchOutsideBraces(declaration: String) -> NSString? {
     return NSString(string: declaration).substringWithRange(outsideBracesMatch.range)
 }
 
+func declarationIsInitializer(declaration: String) -> Bool {
+    return !regex("^((.+)?\\s+)?init\\?*\\(.*\\)")
+        .matchesInString(declaration, options: [],
+                         range: NSRange(location: 0, length: declaration.characters.count)).isEmpty
+}
+
 func commentHasBatchedParameters(comment: String) -> Bool {
     return comment.lowercaseString.containsString("- parameters:")
 }
@@ -76,11 +81,17 @@ func commentReturns(comment: String) -> Bool {
 }
 
 func missingReturnDocumentation(declaration: String, comment: String) -> Bool {
+    guard !declarationIsInitializer(declaration) else {
+        return false
+    }
     return declarationReturns(declaration) && !commentReturns(comment)
 }
 
 func superfluousReturnDocumentation(declaration: String, comment: String,
                                     kind: SwiftDeclarationKind) -> Bool {
+    guard !declarationIsInitializer(declaration) else {
+        return false
+    }
     return !declarationReturns(declaration, kind: kind) && commentReturns(comment)
 }