+++ CHANGELOG.md
@@ -6,9 +6,82 @@
 
 ##### Enhancements
 
-* Add `LegacyNSGeometryFunctionsRule` rule. Add `NSSize`, `NSPoint`, and 
+* Add `RedundantNilCoalesingRule` Opt-In rule that warns against `?? nil`.  
+  [Daniel Beard](https://github.com/daniel-beard)
+  [#764](https://github.com/realm/SwiftLint/issues/764)
+
+* Adds `allow_private_set` configuration for the `private_outlet` rule.  
+  [Rohan Dhaimade](https://github.com/HaloZero)
+
+##### Bug Fixes
+
+* Fixed whitespace being added to TODO messages.  
+  [W. Bagdon](https://github.com/wbagdon)
+  [#792](https://github.com/realm/SwiftLint/issues/792)
+
+* Fixed regex bug in Vertical Whitespace Rule by using SourceKitten instead.
+  The rule now enabled by default again (no longer opt-in).  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#772](https://github.com/realm/SwiftLint/issues/772)
+
+* Fixed regex bug in Mark Rule where MARK could not be used with only a hyphen
+  but no descriptive text: `// MARK: -`.  
+  [Ruotger Deecke](https://github.com/roddi)
+  [#778](https://github.com/realm/SwiftLint/issues/778)
+
+* Fixed: Private unit test rule not scoped to test classes.  
+  Fixed: Private unit test rule config is ignored if regex is missing.  
+  [Cristian Filipov](https://github.com/cfilipov)
+  [#786](https://github.com/realm/SwiftLint/issues/786)
+
+## 0.12.0: Vertical Laundry
+
+##### Breaking
+
+* Fixed: SwiftLint assumes paths in the YAML config file are relative to the
+  current directory even when `--path` is passed as an argument.  
+  [Cristian Filipov](https://github.com/cfilipov)
+
+##### Enhancements
+
+* None.
+
+##### Bug Fixes
+
+* Made Vertical Whitespace Rule added in 0.11.2 opt-in due to performance
+  issues.  
+  [JP Simard](https://github.com/jpsim)
+  [#772](https://github.com/realm/SwiftLint/issues/772)
+
+## 0.11.2: Communal Clothesline
+
+This release has seen a phenomenal uptake in community contributions!
+
+##### Breaking
+
+* None.
+
+##### Enhancements
+
+* Add `MarkRule` rule to enforce `// MARK` syntax.  
+  [Krzysztof Rodak](https://github.com/krodak)
+  [#749](https://github.com/realm/SwiftLint/issues/749)
+
+* Add `PrivateOutletRule` Opt-In rule to enforce `@IBOutlet`
+  instance variables to be `private`.  
+  [Olivier Halligon](https://github.com/AliSoftware)
+
+* Add content of the todo statement to message.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#478](https://github.com/realm/SwiftLint/issues/478)
+
+* Add `LegacyNSGeometryFunctionsRule` rule. Add `NSSize`, `NSPoint`, and
   `NSRect` constants and constructors to existing rules.  
   [David Rönnqvist](https://github.com/d-ronnqvist)
+
+* Added Vertical Whitespace Rule.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#548](https://github.com/realm/SwiftLint/issues/548)
 
 * Removed ConditionalBindingCascadeRule.  
   [J. Cheyo Jimenez](https://github.com/masters3d)
@@ -28,13 +101,37 @@
   [bootstraponline](https://github.com/bootstraponline)
   [#689](https://github.com/realm/SwiftLint/issues/689)
 
+* Add rule to check for private unit tests (private unit tests don't get run 
+  by XCTest).  
+  [Cristian Filipov](https://github.com/cfilipov)
+
+* Add configuration for setting a warning threshold.  
+  [woodhamgh](https://github.com/woodhamgh)
+  [696](https://github.com/realm/SwiftLint/issues/696)
+
+* Adds 'ConditionalReturnsOnNewLineRule' rule.  
+  [Rohan Dhaimade](https://github.com/HaloZero)
+  
+* Made `- returns:` doc optional for initializers.  
+  [Mohpor](https://github.com/mohpor)
+  [#557](https://github.com/realm/SwiftLint/issues/557)
+
 ##### Bug Fixes
 
-* Fix LegacyConstructorRule when using variables instead of numbers.  
-  [Sarr Blaise](https://github.com/bsarr007) 
-  [#646](https://github.com/realm/SwiftLint/issues/646) 
+* Fixed CustomRule Regex.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#717](https://github.com/realm/SwiftLint/issues/717)
+  [#726](https://github.com/realm/SwiftLint/issues/726)
 
-* Fix force_unwrapping false positive inside strings.
+* Allow disabling custom rules in code.  
+  [J. Cheyo Jimenez](https://github.com/masters3d)
+  [#515](https://github.com/realm/SwiftLint/issues/515)
+
+* Fix LegacyConstructorRule when using variables instead of numbers.  
+  [Sarr Blaise](https://github.com/bsarr007)
+  [#646](https://github.com/realm/SwiftLint/issues/646)
+
+* Fix force_unwrapping false positive inside strings.  
   [Daniel Beard](https://github.com/daniel-beard)
   [#721](https://github.com/realm/SwiftLint/issues/721)
 
@@ -57,7 +154,7 @@
 ##### Bug Fixes
 
 * Remove extraneous argument label added in LegacyCGGeometryFunctionsRule
-  autocorrect. 
+  autocorrect.  
   [Sarr Blaise](https://github.com/bsarr007)
   [643](https://github.com/realm/SwiftLint/issues/643)
 
@@ -696,7 +793,7 @@
 
 * Add `autocorrect` command to automatically correct certain violations
   (currently only `trailing_newline`, `trailing_semicolon` &
-  `trailing_whitespace`).
+  `trailing_whitespace`).  
   [JP Simard](https://github.com/jpsim)
   [#5](https://github.com/realm/SwiftLint/issues/5)